Put OpenEXR headers into .cpp, deal with zlib
