+++ 3rdparty/CMakeLists.txt
@@ -90,6 +90,8 @@ set(BUILD_TESTING OFF)
 set(PYILMBASE_ENABLE OFF)
 set(OPENEXR_BUILD_UTILS OFF)
 set(OPENEXR_VIEWERS_ENABLE OFF)
+set(ZLIB_INCLUDE_DIR ${THIRD_PARTY_SOURCE_DIR}/zlib) #for openexr
+set(ZLIB_LIBRARY "$<TARGET_OBJECTS:zlib>") #for openexr
 add_subdirectory(openexr openexr EXCLUDE_FROM_ALL)
 set(BUILD_SHARED_LIBS ${_OLD_BUILD_SHARED_LIBS})
 set(BUILD_STATIC_LIBS ${_OLD_BUILD_STATIC_LIBS})
