+++ src/irr/asset/CImageLoaderOpenEXR.cpp
@@ -22,12 +22,111 @@ SOFTWARE.
 
 #ifdef _IRR_COMPILE_WITH_OPENEXR_LOADER_
 
+#include "openexr/IlmBase/Imath/ImathBox.h"
+#include "openexr/OpenEXR/IlmImf/ImfRgbaFile.h"
+#include "openexr/OpenEXR/IlmImf/ImfStringAttribute.h"
+#include "openexr/OpenEXR/IlmImf/ImfMatrixAttribute.h"
+#include "openexr/OpenEXR/IlmImf/ImfArray.h"
+#include <algorithm>
+#include <iostream>
+#include <string>
+
+#include "openexr/OpenEXR/IlmImf/ImfNamespace.h"
+namespace IMF = OPENEXR_IMF_NAMESPACE;
+namespace IMATH = IMATH_NAMESPACE;
+
 namespace irr
 {
 	namespace asset
 	{
 		using namespace IMF;
 		using namespace IMATH;
+
+		class SContext;
+		bool readVersionField(io::IReadFile* _file, SContext& ctx);
+		bool readHeader(const char fileName[], SContext& ctx);
+		void readRgba(const char fileName[], IMF::Array2D<IMF::Rgba>& pixels, int& width, int& height);
+
+		//! A helpful struct for handling OpenEXR layout
+		/*
+			The latest OpenEXR file consists of the following components:
+			- magic number
+			- version field
+			- header
+			- line offset table
+			- scan line blocks
+		*/
+		struct SContext
+		{
+			constexpr static uint32_t magicNumber = 20000630ul; // 0x76, 0x2f, 0x31, 0x01
+
+			struct VersionField
+			{
+				uint32_t mainDataRegisterField = 0ul;      // treated as 2 seperate bit fields, contains some usefull data
+				uint8_t fileFormatVersionNumber = 0;	   // contains current OpenEXR version. It has to be 0 upon initialization!
+				bool doesFileContainLongNames;			   // if set, the maximum length of attribute names, attribute type names and channel names is 255 bytes. Otherwise 31 bytes
+				bool doesItSupportDeepData;		           // if set, there is at least one part which is not a regular scan line image or regular tiled image, so it is a deep format
+
+				struct Compoment
+				{
+					enum CompomentType
+					{
+						SINGLE_PART_FILE,
+						MULTI_PART_FILE
+					};
+
+					enum SinglePartFileCompoments
+					{
+						NONE,
+						SCAN_LINES,
+						TILES,
+						SCAN_LINES_OR_TILES
+					};
+
+					CompomentType type;
+					SinglePartFileCompoments singlePartFileCompomentSubTypes;
+
+				} Compoment;
+
+			} versionField;
+
+			struct Attributes
+			{
+				// The header of every OpenEXR file must contain at least the following attributes
+				//according to https://www.openexr.com/documentation/openexrfilelayout.pdf (page 8)
+				const IMF::Channel* channels = nullptr;
+				const IMF::Compression* compression = nullptr;
+				const IMATH::Box2i* dataWindow = nullptr;
+				const IMATH::Box2i* displayWindow = nullptr;
+				const IMF::LineOrder* lineOrder = nullptr;
+				const float* pixelAspectRatio = nullptr;
+				const IMATH::V2f* screenWindowCenter = nullptr;
+				const float* screenWindowWidth = nullptr;
+
+				// These attributes are required in the header for all multi - part and /or deep data OpenEXR files
+				const std::string* name = nullptr;
+				const std::string* type = nullptr;
+				const int* version = nullptr;
+				const int* chunkCount = nullptr;
+
+				// This attribute is required in the header for all files which contain deep data (deepscanline or deeptile)
+				const int* maxSamplesPerPixel = nullptr;
+
+				// This attribute is required in the header for all files which contain one or more tiles
+				const IMF::TileDescription* tiles = nullptr;
+
+				// This attribute can be used in the header for multi-part files
+				const std::string* view = nullptr;
+
+				// Others not required that can be used by metadata
+				// - none at the moment
+
+			} attributes;
+
+			// core::smart_refctd_dynamic_array<uint32_t> offsetTable; 
+
+			// scan line blocks TODO
+		};
 
 		asset::SAssetBundle CImageLoaderOpenEXR::loadAsset(io::IReadFile* _file, const asset::IAssetLoader::SAssetLoadParams& _params, asset::IAssetLoader::IAssetLoaderOverride* _override, uint32_t _hierarchyLevel)
 		{
@@ -128,7 +227,7 @@ namespace irr
 				return false;
 		}
 
-		void CImageLoaderOpenEXR::readRgba(const char fileName[], Array2D<Rgba>& pixels, int& width, int& height)
+		void readRgba(const char fileName[], Array2D<Rgba>& pixels, int& width, int& height)
 		{
 			RgbaInputFile file(fileName);
 			Box2i dw = file.dataWindow();
@@ -139,7 +238,7 @@ namespace irr
 			file.readPixels(dw.min.y, dw.max.y);
 		}
 
-		bool CImageLoaderOpenEXR::readVersionField(io::IReadFile* _file, CImageLoaderOpenEXR::SContext& ctx)
+		bool readVersionField(io::IReadFile* _file, SContext& ctx)
 		{
 			RgbaInputFile file(_file->getFileName().c_str());
 			auto& versionField = ctx.versionField;
@@ -181,7 +280,7 @@ namespace irr
 			return true;
 		}
 
-		bool CImageLoaderOpenEXR::readHeader(const char fileName[], CImageLoaderOpenEXR::SContext& ctx)
+		bool readHeader(const char fileName[], SContext& ctx)
 		{
 			RgbaInputFile file(fileName);
 			auto& attribs = ctx.attributes;
